AllowAutoRedirect to True
The new way Avsim downloads files is no longer using the AVSIM_DOWNLOAD_URL_FORMAT, instead, that link returns a ResponseHeader that is supposed to give the browser the location of the file, and download it from there. Now, when the DownloadDataAsync downloads from the downloadUrl, it will redirect to the new location of the file, and save that to the byte array instead. This fixes the error of nothing being downloaded. Also, the filename is not in the headers of the first link, but in the redirect, so this also fixes the "filename not found in response headers." issue.